+++ src/testing/dbTest.ts
@@ -11,31 +11,75 @@ import {
 } from './test.model'
 import { deepFreeze } from './test.util'
 
+export interface CommonDBImplementationFeatures {
+  /**
+   * All querying functionality.
+   */
+  querying?: boolean
+
+  dbQueryFilter?: boolean
+  dbQueryFilterIn?: boolean
+  dbQueryOrder?: boolean
+  dbQuerySelectFields?: boolean
+
+  createTable?: boolean
+  tableSchemas?: boolean
+
+  /**
+   * Queries should return fresh results immediately.
+   * Datastore is the one known to NOT have strong consistency for queries (not for getById though).
+   */
+  strongConsistency?: boolean
+
+  streaming?: boolean
+}
+
 /**
  * All options default to `false`.
  */
-export interface CommonDBTestOptions {
-  allowQueryUnsorted?: boolean
-  allowGetByIdsUnsorted?: boolean
-  allowStreamQueryToBeUnsorted?: boolean
-
+export interface CommonDBImplementationQuirks {
   /**
    * Applicable to e.g Datastore.
    * Time in milliseconds to wait for eventual consistency to propagate.
    */
   eventualConsistencyDelay?: number
 
+  /**
+   * Example: airtableId
+   */
   allowExtraPropertiesInResponse?: boolean
+
+  /**
+   * Example: AirtableDB
+   */
   allowBooleansAsUndefined?: boolean
 }
 
-export function runCommonDBTest(db: CommonDB, opt: CommonDBTestOptions = {}): void {
+/**
+ * All unclaimed features will default to 'true'
+ */
+export function runCommonDBTest(
+  db: CommonDB,
+  features: CommonDBImplementationFeatures = {},
+  quirks: CommonDBImplementationQuirks = {},
+): void {
   const {
-    allowQueryUnsorted,
-    allowGetByIdsUnsorted,
-    allowStreamQueryToBeUnsorted,
-    eventualConsistencyDelay,
-  } = opt
+    querying = true,
+    tableSchemas = true,
+    createTable = true,
+    dbQueryFilter = true,
+    // dbQueryFilterIn = true,
+    dbQueryOrder = true,
+    dbQuerySelectFields = true,
+    streaming = true,
+    strongConsistency = true,
+  } = features
+
+  const {
+    // allowExtraPropertiesInResponse,
+    // allowBooleansAsUndefined,
+  } = quirks
+  const eventualConsistencyDelay = !strongConsistency && quirks.eventualConsistencyDelay
 
   const items = createTestItemsDBM(3)
   deepFreeze(items)
@@ -48,25 +92,29 @@ export function runCommonDBTest(db: CommonDB, opt: CommonDBTestOptions = {}): vo
   })
 
   // CREATE TABLE, DROP
-  test('createTable, dropIfExists=true', async () => {
-    await db.createTable(getTestItemSchema(), { dropIfExists: true })
-  })
+  if (createTable) {
+    test('createTable, dropIfExists=true', async () => {
+      await db.createTable(getTestItemSchema(), { dropIfExists: true })
+    })
+  }
 
-  // DELETE ALL initially
-  test('deleteByIds test items', async () => {
-    const { records } = await db.runQuery(queryAll().select([]))
-    await db.deleteByIds(
-      TEST_TABLE,
-      records.map(i => i.id),
-    )
-  })
+  if (querying) {
+    // DELETE ALL initially
+    test('deleteByIds test items', async () => {
+      const { records } = await db.runQuery(queryAll().select([]))
+      await db.deleteByIds(
+        TEST_TABLE,
+        records.map(i => i.id),
+      )
+    })
 
-  // QUERY empty
-  test('runQuery(all), runQueryCount should return empty', async () => {
-    if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
-    expect((await db.runQuery(queryAll())).records).toEqual([])
-    expect(await db.runQueryCount(queryAll())).toEqual(0)
-  })
+    // QUERY empty
+    test('runQuery(all), runQueryCount should return empty', async () => {
+      if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
+      expect((await db.runQuery(queryAll())).records).toEqual([])
+      expect(await db.runQueryCount(queryAll())).toEqual(0)
+    })
+  }
 
   // GET empty
   test('getByIds(item1.id) should return empty', async () => {
@@ -90,101 +138,120 @@ export function runCommonDBTest(db: CommonDB, opt: CommonDBTestOptions = {}): vo
 
   // GET not empty
   test('getByIds all items', async () => {
-    let records = await db.getByIds<TestItemDBM>(TEST_TABLE, items.map(i => i.id).concat('abcd'))
-    if (allowGetByIdsUnsorted) records = _sortBy(records, 'id')
-    expectMatch(items, records, opt)
+    const records = await db.getByIds<TestItemDBM>(TEST_TABLE, items.map(i => i.id).concat('abcd'))
+    // if (allowGetByIdsUnsorted) records = _sortBy(records, 'id')
+    expectMatch(items, records, quirks)
   })
 
   // QUERY
-  test('runQuery(all) should return all items', async () => {
-    if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
-    let { records } = await db.runQuery(queryAll())
-    if (allowQueryUnsorted) records = _sortBy(records, 'id')
-    expectMatch(items, records, opt)
-  })
+  if (querying) {
+    test('runQuery(all) should return all items', async () => {
+      if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
+      let { records } = await db.runQuery(queryAll())
+      if (!dbQueryOrder) records = _sortBy(records, 'id')
+      expectMatch(items, records, quirks)
+    })
 
-  test('query even=true', async () => {
-    const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).filter('even', '=', true)
-    let { records } = await db.runQuery(q)
-    if (allowQueryUnsorted) records = _sortBy(records, 'id')
-    expectMatch(
-      items.filter(i => i.even),
-      records,
-      opt,
-    )
-  })
+    if (dbQueryFilter) {
+      test('query even=true', async () => {
+        const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).filter('even', '=', true)
+        let { records } = await db.runQuery(q)
+        if (!dbQueryOrder) records = _sortBy(records, 'id')
+        expectMatch(
+          items.filter(i => i.even),
+          records,
+          quirks,
+        )
+      })
+    }
 
-  if (!allowQueryUnsorted) {
-    test('query order by k1 desc', async () => {
-      const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).order('k1', true)
-      const { records } = await db.runQuery(q)
-      expect(records).toEqual([...items].reverse())
+    if (dbQueryOrder) {
+      test('query order by k1 desc', async () => {
+        const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).order('k1', true)
+        const { records } = await db.runQuery(q)
+        expect(records).toEqual([...items].reverse())
+      })
+    }
+
+    if (dbQuerySelectFields) {
+      test('projection query with only ids', async () => {
+        const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).select([])
+        let { records } = await db.runQuery(q)
+        if (!dbQueryOrder) records = _sortBy(records, 'id')
+        expectMatch(
+          items.map(item => _pick(item, ['id'])),
+          records,
+          quirks,
+        )
+      })
+    }
+
+    test('runQueryCount should return 3', async () => {
+      expect(await db.runQueryCount(new DBQuery(TEST_TABLE))).toBe(3)
     })
   }
 
-  test('projection query with only ids', async () => {
-    const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).select([])
-    let { records } = await db.runQuery(q)
-    if (allowQueryUnsorted) records = _sortBy(records, 'id')
-    expectMatch(
-      items.map(item => _pick(item, ['id'])),
-      records,
-      opt,
-    )
-  })
-
-  test('runQueryCount should return 3', async () => {
-    expect(await db.runQueryCount(new DBQuery(TEST_TABLE))).toBe(3)
-  })
-
   // STREAM
-  test('streamQuery all', async () => {
-    let records = await streamMapToArray(db.streamQuery(queryAll()))
+  if (streaming) {
+    test('streamQuery all', async () => {
+      let records = await streamMapToArray(db.streamQuery(queryAll()))
 
-    if (allowStreamQueryToBeUnsorted) records = _sortBy(records, 'id')
-    expectMatch(items, records, opt)
-  })
+      if (!dbQueryOrder) records = _sortBy(records, 'id')
+      expectMatch(items, records, quirks)
+    })
+  }
 
   // getTables
-  test('getTables, getTableSchema', async () => {
+  test('getTables, getTableSchema (if supported)', async () => {
     const tables = await db.getTables()
     console.log({ tables })
-    await pMap(tables, async table => {
-      const schema = await db.getTableSchema(table)
-      console.log(schema)
-      expect(schema).toBeDefined()
-    })
+
+    if (tableSchemas) {
+      await pMap(tables, async table => {
+        const schema = await db.getTableSchema(table)
+        console.log(schema)
+        expect(schema).toBeDefined()
+      })
+    }
   })
 
   // DELETE BY
-  test('deleteByQuery even=false', async () => {
-    const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).filter('even', '=', false)
-    const deleted = await db.deleteByQuery(q)
-    expect(deleted).toBe(items.filter(item => !item.even).length)
+  if (querying && dbQueryFilter) {
+    test('deleteByQuery even=false', async () => {
+      const q = new DBQuery<TestItemBM, TestItemDBM>(TEST_TABLE).filter('even', '=', false)
+      const deleted = await db.deleteByQuery(q)
+      expect(deleted).toBe(items.filter(item => !item.even).length)
 
-    if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
+      if (eventualConsistencyDelay) await pDelay(eventualConsistencyDelay)
 
-    expect(await db.runQueryCount(queryAll())).toBe(1)
-  })
+      expect(await db.runQueryCount(queryAll())).toBe(1)
+    })
+  }
 
-  test('cleanup', async () => {
-    // CLEAN UP
-    const { records } = await db.runQuery(queryAll().select([]))
-    await db.deleteByIds(
-      TEST_TABLE,
-      records.map(i => i.id),
-    )
-  })
+  if (querying) {
+    test('cleanup', async () => {
+      // CLEAN UP
+      const { records } = await db.runQuery(queryAll().select([]))
+      await db.deleteByIds(
+        TEST_TABLE,
+        records.map(i => i.id),
+      )
+    })
+  }
 }
 
-export function expectMatch(expected: any, actual: any, opt: CommonDBTestOptions): void {
-  if (opt.allowBooleansAsUndefined) {
+export function expectMatch(
+  expected: any,
+  actual: any,
+  quirks: CommonDBImplementationQuirks,
+): void {
+  if (quirks.allowBooleansAsUndefined) {
     expected = (Array.isArray(expected) ? expected : [expected]).map(r =>
       filterObject(r, (_k, v) => v !== false),
     )
   }
 
-  if (opt.allowExtraPropertiesInResponse) {
+  if (quirks.allowExtraPropertiesInResponse) {
     expect(actual).toMatchObject(expected)
   } else {
     expect(actual).toEqual(expected)
